+++ tests/kwargs/__init__.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Form(BaseModel):
+    name: str
+    age: int
+    programmer: bool
