+++ tests/kwargs/test_path_params.py
@@ -1,0 +1,97 @@
+from uuid import uuid1, uuid4
+
+import pytest
+from pydantic import UUID4
+from starlette.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
+
+from starlite import (
+    ImproperlyConfiguredException,
+    Parameter,
+    Starlite,
+    create_test_client,
+    get,
+)
+
+
+@pytest.mark.parametrize(
+    "params_dict,should_raise",
+    [
+        (
+            {
+                "version": 1.0,
+                "service_id": 1,
+                "user_id": "abc",
+                "order_id": str(uuid4()),
+            },
+            False,
+        ),
+        (
+            {
+                "version": 4.1,
+                "service_id": 1,
+                "user_id": "abc",
+                "order_id": str(uuid4()),
+            },
+            True,
+        ),
+        (
+            {
+                "version": 0.2,
+                "service_id": 101,
+                "user_id": "abc",
+                "order_id": str(uuid4()),
+            },
+            True,
+        ),
+        (
+            {
+                "version": 0.2,
+                "service_id": 1,
+                "user_id": "abcdefghijklm",
+                "order_id": str(uuid4()),
+            },
+            True,
+        ),
+        (
+            {
+                "version": 0.2,
+                "service_id": 1,
+                "user_id": "abc",
+                "order_id": str(uuid1()),
+            },
+            True,
+        ),
+    ],
+)
+def test_path_params(params_dict: dict, should_raise: bool):
+    test_path = "{version:float}/{service_id:int}/{user_id:str}/{order_id:uuid}"
+
+    @get(path=test_path)
+    def test_method(
+        order_id: UUID4,
+        version: float = Parameter(gt=0.1, le=4.0),
+        service_id: int = Parameter(gt=0, le=100),
+        user_id: str = Parameter(min_length=1, max_length=10),
+    ) -> None:
+        assert version
+        assert service_id
+        assert user_id
+        assert order_id
+
+    with create_test_client(test_method) as client:
+        response = client.get(
+            f"{params_dict['version']}/{params_dict['service_id']}/{params_dict['user_id']}/{params_dict['order_id']}"
+        )
+        if should_raise:
+            assert response.status_code == HTTP_400_BAD_REQUEST
+        else:
+            assert response.status_code == HTTP_200_OK
+
+
+def test_path_param_validation():
+    @get(path="/{param}")
+    def test_method() -> None:
+        pass
+
+    with pytest.raises(ImproperlyConfiguredException):
+        Starlite(route_handlers=[test_method])
