+++ tests/kwargs/test_url_encoded_data.py
@@ -1,0 +1,19 @@
+from starlette.status import HTTP_201_CREATED
+
+from starlite import Body, RequestEncodingType, create_test_client, post
+from tests.kwargs import Form
+
+
+def test_request_body_url_encoded():
+    body = Body(media_type=RequestEncodingType.URL_ENCODED)
+
+    test_path = "/test"
+    data = Form(name="Moishe Zuchmir", age=30, programmer=True).dict()
+
+    @post(path=test_path)
+    def test_method(data: Form = body) -> None:
+        assert isinstance(data, Form)
+
+    with create_test_client(test_method) as client:
+        response = client.post(test_path, data=data)
+        assert response.status_code == HTTP_201_CREATED
